Created isListValid method with all helper methods
Helper methods check if a job is valid compared to all parallel jobs
running with it, and against all previously running jobs. Not
implemented yet: matching these helper methods with job ids to make sure
each job is depending on the current job. For now, we assume that is
already is.